requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
aiosqlite>=0.20
argon2-cffi>=23.1
webauthn>=2.0
itsdangerous>=2.1
python-multipart>=0.0.9
jinja2>=3.1
httpx>=0.27
Pillow>=10.0
pydantic-settings>=2.0
